+++ source/appModules/winword.py
@@ -1,0 +1,43 @@
+# appModules/winword.py
+# A part of NonVisual Desktop Access (NVDA)
+# This file is covered by the GNU General Public License.
+# See the file COPYING for more details.
+# Copyright (C) 2020 NV Access Limited.
+
+""" App module for Microsoft Word.
+Word and Outlook share a lot of code and components. This app module gathers the code that is relevant for
+Microsoft Word only.
+"""
+
+import appModuleHandler
+from scriptHandler import script
+import ui
+from NVDAObjects.IAccessible.winword import WordDocument
+
+
+class AppModule(appModuleHandler.AppModule):
+
+	def chooseNVDAObjectOverlayClasses(self, obj, clsList):
+		if WordDocument in clsList:
+			clsList.insert(0, WinwordWordDocument)
+
+
+class WinwordWordDocument(WordDocument):
+
+	@script(gesture="kb:control+shift+e")
+	def script_toggleChangeTracking(self, gesture):
+		if not self.WinwordDocumentObject:
+			# We cannot fetch the Word object model, so we therefore cannot report the status change.
+			# The object model may be unavailable because it's within Windows Defender Application Guard.
+			# In this case, just let the gesture through and don't report anything.
+			return gesture.send()
+		val = self._WaitForValueChangeForAction(
+			lambda: gesture.send(),
+			lambda: self.WinwordDocumentObject.TrackRevisions
+		)
+		if val:
+			# Translators: a message when toggling change tracking in Microsoft word
+			ui.message(_("Change tracking on"))
+		else:
+			# Translators: a message when toggling change tracking in Microsoft word
+			ui.message(_("Change tracking off"))
